pyproject.toml
[project]
name = "calnio"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiocaldav>=0.5.1",
    "aiosmtplib==3.0.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "apscheduler>=3.11.1",
    "asyncpg==0.30.0",
    "authx==1.4.3",
    "bcrypt==4.0.1",
    "caldav>=2.1.2",
    "celery>=5.6.0",
    "click==8.2.1",
    "colorlog>=6.10.1",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi==0.115.14",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httpx>=0.28.1",
    "icalendar>=6.3.2",
    "idna==3.10",
    "jinja2==3.1.6",
    "litestar>=2.18.0",
    "loguru>=0.7.3",
    "markupsafe==3.0.2",
    "notion-client>=2.7.0",
    "passlib[bcrypt]==1.7.4",
    "prometheus-client>=0.23.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psycopg==3.2.9",
    "psycopg2-binary==2.9.10",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pytest>=9.0.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-multipart==0.0.20",
    "recurring-ical-events>=3.8.0",
    "redis>=7.1.0",
    "sniffio==1.3.1",
    "sqladmin==0.21.0",
    "sqlalchemy==2.0.41",
    "starlette==0.46.2",
    "tabulate==0.9.0",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "uvicorn==0.35.0",
    "wtforms==3.1.2",
]
